Fix './run install' output; remove log level stuff
I fixed the './run install' command so the output isn't so verbose.

I misunderstood the logging stuff initially. I removed the logging stuff so it can be handled from main.rs instead of from run. One way to implement it would be to read the environment variable from within main.rs (similar to how GITHUB_TOKEN is read) and use that to determine the log verbosity.